requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.27.1
webdriver-manager==4.0.2
